LayoutTests/imported/w3c/web-platform-tests/screen-orientation/nested-documents.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<body>
  <script type="module">
    import {
      attachIframe,
      makeCleanup,
      getOppositeOrientation,
    } from "./resources/orientation-utils.js";

    promise_test(async (t) => {
      const iframe = await attachIframe();
      const iframeWin = iframe.contentWindow;

      // Go full screen
      await test_driver.bless("request full screen", null, iframeWin);
      await iframe.contentDocument.body.requestFullscreen();

      // Lock the orientation from the iframe
      const opposite = getOppositeOrientation();

      const topPromise = window.screen.orientation.lock(opposite);
      // Causes topPromise to abort
      const iframePromise = iframeWin.screen.orientation.lock(opposite);
      // Causes iframePromise to abort
      const exitFullScreenPromise = iframe.contentDocument.exitFullscreen();

      await promise_rejects_dom(t, "AbortError", topPromise);
      await promise_rejects_dom(
        t,
        "AbortError",
        iframeWin.DOMException,
        iframePromise
      );
      await exitFullScreenPromise;
    }, "Requesting orientation lock from one document cancels the lock request from another document");

    promise_test(async (t) => {
      // Create 2 nested iframes
      const outerIframe = await attachIframe({ src: `about:blank#1` });
      const innerIframe = await attachIframe({
        context: outerIframe.contentWindow,
        src: `about:blank#2`,
      });

      const iframes = [outerIframe, innerIframe];

      // Go full screen on the innerIframe
      await test_driver.bless(
        "request full screen",
        null,
        innerIframe.contentWindow
      );
      await innerIframe.contentDocument.body.requestFullscreen();

      const opposite = getOppositeOrientation();

      // Each iframe tries to lock the orientation,
      const requestsToLock = iframes.map((iframe) => {
        return {
          promise: iframe.contentWindow.screen.orientation.lock(opposite),
          context: iframe.contentWindow,
          iframe,
        };
      });

      // But calling lock() from top-level aborts the last promise
      const topPromise = window.screen.orientation.lock(opposite);

      // Check that all promises are rejected with AbortError
      for (const { promise, context, iframe } of requestsToLock) {
        await promise_rejects_dom(
          t,
          "AbortError",
          context.DOMException,
          promise,
          `iframe ${iframe.contentWindow.location}`
        );
      }
      // Finally, top-level promise resolves
      await topPromise;
    }, "The orientation lock from one document affects lock requests from other documents");
  </script>
</body>
